+++ app/views/projects/shared/_new_projects.html.erb
@@ -5,7 +5,7 @@
           class: "govuk-button",
           data: {module: "govuk-button"} %>
     <%= link_to t("conversion_project.involuntary.new.title"),
-          conversion_involuntary_new_path,
+          new_conversions_involuntary_project_path,
           class: "govuk-button",
           data: {module: "govuk-button"} %>
   </div>
