+++ spec/features/regional_delivery_officers_can_create_new_projects_spec.rb
@@ -17,7 +17,7 @@ RSpec.feature "Regional delivery officers can create new projects" do
       visit projects_path
 
       expect(page)
-        .to have_link I18n.t("conversion_project.involuntary.new.title"), href: conversion_involuntary_new_path
+        .to have_link I18n.t("conversion_project.involuntary.new.title"), href: new_conversions_involuntary_project_path
     end
   end
 end
