+++ spec/requests/conversions/involuntary/projects_contoller_spec.rb
@@ -2,6 +2,21 @@ require "rails_helper"
 
 RSpec.describe Conversions::Involuntary::ProjectsController do
   let(:regional_delivery_officer) { create(:user, :regional_delivery_officer) }
+  let(:create_path) { conversions_involuntary_projects_path }
+  let(:workflow_root) { Conversion::Involuntary::Details::WORKFLOW_PATH }
+  let(:form_class) { Conversion::Involuntary::CreateProjectForm }
+  let(:project_form) { build(:create_involuntary_project_form) }
+  let(:project_form_params_key) { "conversion_involuntary_create_project_form" }
+  let(:project_form_params) {
+    attributes_for(:create_involuntary_project_form,
+      "provisional_conversion_date(3i)": "1",
+      "provisional_conversion_date(2i)": "1",
+      "provisional_conversion_date(1i)": "2030",
+      "advisory_board_date(3i)": "1",
+      "advisory_board_date(2i)": "1",
+      "advisory_board_date(1i)": "2022",
+      regional_delivery_officer: nil)
+  }
 
   before do
     establishment = build(:academies_api_establishment)
@@ -10,6 +25,8 @@ RSpec.describe Conversions::Involuntary::ProjectsController do
     mock_successful_api_calls(establishment: establishment, trust: trust)
     sign_in_with(regional_delivery_officer)
   end
+
+  it_behaves_like "a conversion project"
 
   describe "#index" do
     it "lists involuntary conversion projects" do
